Generate Id for GridState

+++ src/main/java/com/dandykong/butter/game/CellState.kt
@@ -1,6 +1,6 @@
 package com.dandykong.butter.game
 
-enum class CellState(val state: Byte) {
+enum class CellState(val state: Int) {
     EMPTY(0),
     MINE(1),
     THEIRS(2)
